+++ app/src/main/java/org/mozilla/tv/firefox/pocket/PocketViewModel.kt
@@ -22,7 +22,6 @@ const val POCKET_VIDEO_COUNT = 20
  */
 class PocketViewModel(
     private val pocketVideoRepo: PocketVideoRepo,
-    private val localeIsEnglish: () -> Boolean,
     pocketRepoCache: PocketRepoCache
 ) : ViewModel() {
 
@@ -55,13 +54,8 @@ class PocketViewModel(
                 is PocketVideoRepo.FeedState.LoadComplete -> State.Feed(repoState.videos)
                 is PocketVideoRepo.FeedState.NoAPIKey -> State.Feed(noKeyPlaceholders)
                 is PocketVideoRepo.FeedState.FetchFailed -> State.Error
+                is PocketVideoRepo.FeedState.Inactive -> State.NotDisplayed
             }
-        }.map {
-            if (localeIsEnglish()) it
-            // We only show the Pocket mega tile if the current language is English.
-            // Otherwise, overwrite any state with NotDisplayed
-            // See issue: https://github.com/mozilla-mobile/firefox-tv/issues/1283
-            else State.NotDisplayed
         }
 
     companion object {