+++ app/src/main/java/org/mozilla/tv/firefox/utils/BuildConfigDerivables.kt
@@ -15,10 +15,11 @@ private const val POCKET_PARAM_API_KEY = "consumer_key"
  *
  * This logic is often simple but noisy, so pulling it out of client code improves readability.
  */
-class BuildConfigDerivables {
+class BuildConfigDerivables(localeIsEnglish: () -> Boolean) {
     @Suppress("UselessCallOnNotNull") // Values of BuildConfig can change but the compiler doesn't know that
     val initialPocketRepoState = when {
         BuildConfig.POCKET_KEY.isNullOrEmpty() -> PocketVideoRepo.FeedState.NoAPIKey
+        !localeIsEnglish.invoke() -> PocketVideoRepo.FeedState.Inactive
         else -> PocketVideoRepo.FeedState.Loading
     }