+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Some web fonts were blocked which caused icons to be missing
 - Pocket requests are no longer made if a valid key is not found
+- Pocket requests are no longer made when the locale is not set to English
 
 ## [3.1] - 2018-11-16
 ### Added